Use less aggressive global gitignore
There were problems with Python build directories being applied in
non-Python projects (for example), which led to files being ignored
when they shouldn't have been. This has happened often enough that it's
not worth it to keep those entries in my global gitignore, even if I do
have a lot of Python projects.

The solution should be to generate gitignore files on a per-project
bases, using a service like gitignore.io.