Reset tmux prefix to C-b
When switching between Qwerty and Dvorak, and between local and remote
servers, it's confusing to remember which shortcut I'm using. C-b isn't
that bad on my principle keyboard layout of Dvorak anyways, so I might
as well switch to using it, since it's the default, and I should never
get confused in different environments from this point forward.