Attempt to fix sharing history between tmux tabs
Supposedly the option that was already set takes care of this case, but
in practice, I've been having issues with zsh in different tabs
remembering commands from other tabs. This is pretty confusing when I
try to use ^P or whatever to run the last command that I ran in the tab
in question. This often happens because tabs generally have different
purposes, such as one tab to edit and one tab to build.